Changes to Window Resize
I've made changes to our JS and CSS files to offer a different method of resizing out application on a webpage. However, next week I will most likely work on making drastic changes to the layout as the connection between the ChatGPT API key and the front end are coming together.